refactor: split webhook delivery out of check.go, emptying the allowlist
check.go was allowlisted past the 800-line ceiling on the grounds that it is a
state machine hardened over 21 review rounds and splitting risks regressing
currency/dedup/webhook logic (trvl#551).

Half of that holds. The webhook half does not: notification delivery is not the
state machine. It is an SSRF guard, a redirect policy, a credential redaction
layer and an HTTP client; it already has its own tests (webhook_test.go,
webhook_log_test.go); and check.go's entire dependency on it is one call to
fireWebhook. The "21 review rounds" argument protects the price-check logic,
which is not what moved.

check.go 826 -> 549, webhook.go 306. Verbatim move, two contiguous blocks: the
guarded outbound client with its CIDR blocks, and the payload/delivery/redaction
helpers. The new file's header states the three security properties together --
non-public destinations refused at dial time, redirects refused rather than
followed so a receiver cannot launder that check, and nothing logged beyond
scheme+host because the URL IS the credential -- which is the actual argument
for the file existing.

.github/loc-allowlist.txt now lists no source file at all. Every entry is a gate
someone decided not to meet, so the list is kept to files where the ceiling does
not apply (CHANGELOG.md, append-only release history) rather than to code that
has not been split yet.

Largest remaining source file is 800 lines exactly.

Verification: internal/watch 298 passed, go build clean, go vet clean, gofmt
clean, all four scripts/ci hygiene checks pass.

trvl#551 can close once this reaches origin/main.
